requirements.txt
argon2-cffi==23.1.0
flask==3.0.3
Flask_SQLAlchemy==3.1.1
floodgate==2.0.0
gunicorn==22.0.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
pyjwt[crypto]==2.8.0
requests==2.32.3
SQLAlchemy==2.0.31
